Delete schedule and template properly

+++ servers/core/Handlers/Templates/TemplateDelete.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using System.Threading.Tasks;
 using Core.Utility;
 using Microsoft.AspNetCore.Http;
@@ -16,6 +17,14 @@ public static partial class TemplateHandler
             return Results.NotFound("Template not found");
         }
 
+        var schedules = await db.NotificationSchedules.Where(s => s.TemplateId == id).ToListAsync();
+
+        if (schedules.Any())
+        {
+            return Results.NotFound("There are schedules associated with this template");
+        }
+
+        template.ActiveStudents.Clear();
         db.NotificationTemplates.Remove(template);
 
         await db.SaveChangesAsync();
